+++ project/api/index.py
@@ -1,8 +1,8 @@
-import os
-from django.core.asgi import get_asgi_application
+from project.wsgi import application
 from whitenoise import WhiteNoise
+import os
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'project.settings')
-
-application = get_asgi_application()
-app = WhiteNoise(application, root=os.path.join(os.path.dirname(__file__), "..", "staticfiles"))
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+global_static_path = os.path.join(BASE_DIR, "static")
+application = WhiteNoise(application, root=global_static_path, prefix='static/')
+app = application